Adding Senior Mod Type
Splitting the moderator type into 2 types. The goal is creating a new basic moderator type where new moderators have access only to tools needed to complete basic jobs. More complex or dangerous tools used by senior moderators have been moved into the senior moderator class. The inhertance has been changed so that all senior moderators will by default have access to all forum moderator tools. New admin tools have been included to manage senior moderator status, and senior mods are displayed on the status page in the admincp.
