+++ admin/adminActionsForms.php
@@ -22,6 +22,7 @@ defined('IN_CODE') or die('This script can not be run by itself.');
 
 require_once(l_r('admin/adminActions.php'));
 require_once(l_r('admin/adminActionsForum.php'));
+require_once(l_r('admin/adminActionsSeniorMod.php'));
 require_once(l_r('admin/adminActionsRestricted.php'));
 require_once(l_r('admin/adminActionsTD.php'));
 
@@ -111,7 +112,8 @@ class adminActionsForms
 
 	private static function isActionDangerous($actionCode)
 	{
-		if ( method_exists('adminActions', $actionCode.'Confirm')||method_exists('adminActionsRestricted', $actionCode.'Confirm') )
+		if ( method_exists('adminActions', $actionCode.'Confirm')|| method_exists('adminActionsRestricted', $actionCode.'Confirm') 
+			|| method_exists('adminActionsSeniorMods', $actionCode.'Confirm'))
 			return false;
 		else
 			return true;
@@ -363,6 +365,8 @@ else
 
 	if ( $User->type['Admin'] )
 		$adminActions = new adminActionsRestricted();
+	elseif ( $User->type['SeniorMod'] )
+	 	$adminActions = new adminActionsSeniorMod();
 	elseif ( $User->type['ForumModerator'] )
 		$adminActions = new adminActionsForum();
 	else
