+++ install/1.64-1.65/readme.txt
@@ -1,0 +1,3 @@
+Changelog
+---------
+* Adding a new type for a senior mod
